+++ src/views/index.handlebars
@@ -1,4 +1,4 @@
- {{#if loggedIn}}
+ {{#if authed}}
     <h1>You are logged in</h1>
 {{else}}
     {{> login}}
